[Refactor] 폴더 구조 통일 (#192)
* refactor: 폴더 정리

* refactor: query를 hook으로 분리

* chore: useEffect 의존성 수정

* fix: 이메일 인증 버튼 클릭시 기존 인증 번호 지우기
