+++ src/common/components/Input/components/InputButton/types.ts
@@ -1,0 +1,6 @@
+export interface InputButtonProps {
+  text: string;
+  isLoading?: boolean;
+  disabled?: boolean;
+  onClick: () => void;
+}
