+++ src/common/components/Input/components/Timer/utils/formatTimer.ts
