+++ src/common/components/Input/index.tsx
@@ -1,5 +1,5 @@
-export { default as Description } from './Description';
-export { default as InputButton } from './InputButton';
-export { default as InputLine } from './InputLine';
-export { TextBox } from './TextBox';
-export { default as Timer } from './Timer';
+export { default as Description } from './components/Description';
+export { default as InputButton } from './components/InputButton';
+export { default as InputLine } from './components/InputLine';
+export { TextBox } from './components/TextBox';
+export { default as Timer } from './components/Timer';
